backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.1.2
python-multipart>=0.0.6
aiofiles>=23.2.1
yt-dlp>=2023.11.16
aiomysql>=0.2.0
redis>=5.0.1
httpx>=0.25.2
itsdangerous==2.2.0
emails==0.6
stripe==7.11.0
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.4
tenacity==8.2.3
celery>=5.3.5
email-validator>=2.1.0
aiohttp>=3.9.0
# Добавленные библиотеки
google-auth>=2.23.0
authlib>=1.2.1
ffmpeg-python>=0.2.0
yookassa>=2.4.0
pymysql>=1.1.0
instaloader>=4.10.1
jinja2>=3.1.2
pillow>=10.0.0 